# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List all proposals
      responses:
        "200":
          description: A list of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update a proposal
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalUpdate"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /proposals/{id}/convert:
    post:
      operationId: convertProposal
      tags: [proposals]
      summary: Convert an accepted quote into a job and a draft invoice
      description: |
        Creates a job (with per-job materials and a material-readiness record)
        and a draft invoice from an accepted quote, carrying over line items,
        salesperson, totals, and the deposit. Idempotent — returns the existing
        job and invoice if the quote was already converted.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Existing job and invoice (already converted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertProposalResult"
        "201":
          description: Created job and invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertProposalResult"
        "400":
          description: Quote is not in an acceptable state to convert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    JobStatus:
      type: string
      enum:
        - New Lead
        - Estimate Scheduled
        - Estimate Completed
        - Proposal Sent
        - Approved
        - Material Ordered
        - Material Received
        - Scheduled
        - In Progress
        - Final Walkthrough
        - Completed
        - Invoiced
    PriorityLevel:
      type: string
      enum: [Low, Medium, High]
    RiskLevel:
      type: string
      enum: [Low, Medium, High]
    CrewAssigned:
      type: string
      enum: [Crew A, Crew B, Crew C, Crew D, Unassigned]
    FlooringType:
      type: string
      enum:
        - Carpet
        - Hardwood
        - Tile
        - Laminate
        - Luxury Vinyl Plank (LVP)
        - Luxury Vinyl Tile (LVT)
        - Waterproof Flooring
        - Commercial Carpet Tile
        - Commercial LVT
    MaterialStatusBadge:
      type: string
      enum:
        - Ordered
        - In Transit
        - Received
        - Delayed
        - Damaged
        - Missing Items
    ProposalStatus:
      type: string
      enum: [Draft, Sent, Viewed, Accepted, Declined]
    DepositType:
      type: string
      enum: [none, percent, amount]
    InvoiceStatus:
      type: string
      enum: [Draft, Sent, Paid, Overdue]
    InvoiceLineCategory:
      type: string
      enum: [Labor, Materials, Add-on]
...
    Room:
      type: object
      required: [id, name, length, width]
      properties:
        id:
          type: string
        name:
          type: string
        length:
          type: number
        width:
          type: number
        scope:
          type: string
    StageEvent:
      type: object
      required: [stage, at]
      properties:
        stage:
          $ref: "#/components/schemas/JobStatus"
        at:
          type: string
...
    Job:
      type: object
      required:
        - id
        - jobNumber
        - customerName
        - phone
        - email
        - address
        - city
        - flooringType
        - rooms
        - scopeOfWork
        - squareFootage
        - crewAssigned
        - materialStatus
        - laborEstimate
        - estLaborHours
        - materialEstimate
        - estRevenue
        - estGrossProfit
        - grossMarginPct
        - notes
        - priorityLevel
        - riskLevel
        - status
        - stageHistory
        - shareToken
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        jobNumber:
          type: string
        customerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        scopeOfWork:
          type: string
        squareFootage:
          type: number
        crewAssigned:
          $ref: "#/components/schemas/CrewAssigned"
        salespersonId:
          type: string
          nullable: true
        estStartDate:
          type: string
        estCompletionDate:
          type: string
        materialStatus:
          $ref: "#/components/schemas/MaterialStatusBadge"
        laborEstimate:
          type: number
        estLaborHours:
          type: number
        materialEstimate:
          type: number
        estRevenue:
          type: number
        estGrossProfit:
          type: number
        grossMarginPct:
          type: number
        actualRevenue:
          type: number
        actualLaborCost:
          type: number
        actualMaterialCost:
          type: number
        actualAddOnCost:
          type: number
        notes:
          type: string
        priorityLevel:
          $ref: "#/components/schemas/PriorityLevel"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        status:
          $ref: "#/components/schemas/JobStatus"
        stageHistory:
          type: array
          items:
            $ref: "#/components/schemas/StageEvent"
        shareToken:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Proposal:
      type: object
      required:
        - id
        - customerName
        - projectLocation
        - flooringType
        - roomList
        - totalSqFt
        - scopeOfWork
        - estimatedPrice
        - expectedTimeline
        - materialAssumptions
        - exclusions
        - warrantyNote
        - depositType
        - depositValue
        - paymentTerms
        - status
        - shareToken
        - signature
        - createdAt
      properties:
        id:
          type: string
        jobId:
          type: string
        customerName:
          type: string
        projectLocation:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        roomList:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ProposalLineItem"
        salespersonId:
          type: string
        totalSqFt:
          type: number
        scopeOfWork:
          type: string
        estimatedPrice:
          type: number
        expectedTimeline:
          type: string
        materialAssumptions:
          type: string
        exclusions:
          type: string
        warrantyNote:
          type: string
        depositType:
          $ref: "#/components/schemas/DepositType"
        depositValue:
          type: number
        paymentTerms:
          type: string
        status:
          $ref: "#/components/schemas/ProposalStatus"
        shareToken:
          type: string
        sentAt:
          type: string
        viewedAt:
          type: string
        acceptedAt:
          type: string
        declinedAt:
          type: string
        signature:
          type: string
        convertedJobId:
          type: string
          nullable: true
        convertedInvoiceId:
          type: string
          nullable: true
        createdAt:
          type: string
    ProposalLineItem:
      type: object
      required:
        - id
        - name
        - unit
        - quantity
        - unitPrice
      properties:
        id:
          type: string
        productId:
          type: string
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        unit:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
    ProposalInput:
      type: object
      required:
        - customerName
        - projectLocation
        - flooringType
        - roomList
        - totalSqFt
        - scopeOfWork
        - estimatedPrice
        - expectedTimeline
        - materialAssumptions
        - exclusions
        - warrantyNote
        - status
      properties:
        jobId:
          type: string
        customerName:
          type: string
        projectLocation:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        roomList:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ProposalLineItem"
        salespersonId:
          type: string
        totalSqFt:
          type: number
        scopeOfWork:
          type: string
        estimatedPrice:
          type: number
        expectedTimeline:
          type: string
        materialAssumptions:
          type: string
        exclusions:
          type: string
        warrantyNote:
          type: string
        depositType:
          $ref: "#/components/schemas/DepositType"
        depositValue:
          type: number
        paymentTerms:
          type: string
        status:
          $ref: "#/components/schemas/ProposalStatus"
    ProposalUpdate:
      type: object
      properties:
        customerName:
          type: string
        projectLocation:
          type: string
        scopeOfWork:
          type: string
        expectedTimeline:
          type: string
        materialAssumptions:
          type: string
        exclusions:
          type: string
        warrantyNote:
          type: string
        depositType:
          $ref: "#/components/schemas/DepositType"
        depositValue:
          type: number
        paymentTerms:
          type: string
        status:
          $ref: "#/components/schemas/ProposalStatus"
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ProposalLineItem"
        salespersonId:
          type: string
          nullable: true
...
    ConvertProposalResult:
      type: object
      required: [job, invoice, materialsCreated]
      properties:
        job:
          $ref: "#/components/schemas/Job"
        invoice:
          $ref: "#/components/schemas/Invoice"
        materialsCreated:
          type: integer
...
    InvoiceLineItem:
      type: object
      required: [id, description, category, quantity, unitPrice]
      properties:
        id:
          type: string
        description:
          type: string
        category:
          $ref: "#/components/schemas/InvoiceLineCategory"
        quantity:
          type: number
        unitPrice:
          type: number
    Invoice:
      type: object
      required:
        - id
        - invoiceNumber
        - jobId
        - jobNumber
        - customerName
        - lineItems
        - subtotal
        - taxableAmount
        - taxAmount
        - discountAmount
        - total
        - depositAmount
        - paidAmount
        - balanceAmount
        - refundedAmount
        - taxCode
        - paymentReference
        - status
        - issueDate
        - dueDate
        - notes
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        jobId:
          type: string
        jobNumber:
          type: string
        customerName:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        taxableAmount:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        depositAmount:
          type: number
        paidAmount:
          type: number
        balanceAmount:
          type: number
        refundedAmount:
          type: number
        taxCode:
          type: string
        paymentReference:
          type: string
        paidAt:
          type: string
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        issueDate:
          type: string
        dueDate:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...